backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
youtube-transcript-api>=1.2.3
openai>=1.0.0
python-dotenv>=1.0.0
yt-dlp>=2023.10.0
pydantic>=2.0.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0
